smartdatapay-predev/openai-fineturning-in-saas/src/requirements.txt
python-dotenv
requests
pandas
openai
gradio
